+++ src/Vault.ts
@@ -1,60 +1,9 @@
 import { Account } from './Account';
 import { masterVault as backend } from '@xbacked-dao/xbacked-contracts';
+import { VaultReturnParams, ReachUserVault, UserVaultReturnParams, VaultParameters } from './interfaces';
 /**
  * The Parameters returned from the staate of a contract
  */
-export interface VaultReturnParams {
-  /** @property Fees that have accrued in the contract */
-  accruedFees: number;
-  /** @property The present collateral price in the contract */
-  collateralPrice: number;
-  /** @property Indicator that signifies if the contract is deprecated or not */
-  deprecated: boolean;
-  /** @property The percentage of the accruedFees, sent to the FeeCollector as a reward for distributing fees */
-  feeCollectorFee: number;
-  /** @property Minimum collateral ratio for a vault to be liquidated in micro units */
-  liquidationCollateralRatio: number;
-  /** @property The percentage of the amount liquidated, taken by the contract as fees during liquidation */
-  liquidationFee: number;
-  /** @property The minimum collateral ratio allowed for vault creation, minting, withdrawal in the contract */
-  minimumCollateralRatio: number;
-  /** @property The total amount of debt in the contract */
-  totalVaultDebt: number;
-  /** @property An array of addresses for redeemable vaults */
-  redeemableVaults: any[];
-  /** @property the accrued interest in a vault awaiting distribution via settleInterest */
-  accruedInterest: number;
-  /** @property the interest rate of a vault */
-  interestRate: number;
-}
-
-export interface ReachUserVault {
-  /** @property The collateral ratio for a vault */
-  collateralRatio: number;
-  /** @property The amount of collateral in a vault */
-  collateral: number;
-  /** @property Indicator that signifies if a vault can be liquidated */
-  liquidating: boolean;
-  /** @property The debt in a vault */
-  vaultDebt: number;
-  /** @property Inidicator that signifies if a vault can be redeemed */
-  redeemable: boolean;
-  /** @property the timestamp of the last time interest accrued for a specific vault */
-  lastAccruedInterestTime: number;
-}
-
-export interface UserVaultReturnParams extends ReachUserVault {
-  /** @property Indicator that signifies that a vault exists for a particular address */
-  vaultFound: boolean;
-}
-/**
- * parameters used to instantiate the Contract constructor
- */
-export interface VaultParameters {
-  /**  @property Unique identifier for the contract */
-  id: number;
-  acc?: any;
-}
 /**
  * An abstraction of an Xusd contract
  */
